fix inventory refresh issues?

+++ TCC.Core/Data/GearItem.cs
@@ -7,7 +7,7 @@
         public GearPiece Piece { get; private set; }
 
         public int Enchant { get; private set; }
-        public uint Experience { get; private set; }
+        public long Experience { get; private set; }
         public int MaxExperience => SessionManager.CurrentDatabase.GetItemMaxExp(Id, Enchant);
 
         public double ExperienceFactor
@@ -43,7 +43,7 @@
         }
 
         public string Name => SessionManager.CurrentDatabase.ItemsDatabase.GetItemName(Id);
-        public GearItem(uint id, GearTier t, GearPiece p, int enchant, uint exp)
+        public GearItem(uint id, GearTier t, GearPiece p, int enchant, long exp)
         {
             Id = id;
             Tier = t;
